[EasyBugsnag] Make shutdown strategy consistent + fix notify in laravel queue
